fix the compiler's behaviour with pipes and binary operations
